LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-ABA.sub.https.window-expected.txt

CONSOLE MESSAGE: Blocked access to external URL https://www.localhost:9443/storage-access-api/resources/requestStorageAccess-iframe.https.html?testCase=ABA

Harness Error (TIMEOUT), message = null
